Move measures.dummy_data package under dummy_data
