+++ ehrql/measures/__init__.py
@@ -1,9 +1,9 @@
+from ehrql.dummy_data.measures import DummyMeasuresDataGenerator
 from ehrql.measures.calculate import (
     get_column_specs_for_measures,
     get_measure_results,
 )
 from ehrql.measures.disclosure_control import apply_sdc_to_measure_results
-from ehrql.measures.dummy_data import DummyMeasuresDataGenerator
 from ehrql.measures.measures import INTERVAL, Measures, create_measures
