+++ tests/acceptance/external_studies/mainroute_cancer/analysis/define_static_dataset.py
